Formatting and linting. Run premake, read README!
Introduces formatting with each compilation. Manual formatting via premake4 format.
Introduces linting before each commit. Manual linting via premake4 lint.
Initial formatting and linting.

+++ src/lib/operators/abstract_operator.hpp
@@ -1,33 +1,35 @@
 #pragma once
 
+#include <memory>
+#include <string>
+#include <vector>
+
 #include "../common.hpp"
 #include "../storage/table.hpp"
-
-#include <vector>
 
 namespace opossum {
 
 class abstract_operator {
-public:
-    abstract_operator(const std::shared_ptr<abstract_operator> left = nullptr, const std::shared_ptr<abstract_operator> right = nullptr);
-    abstract_operator(abstract_operator const&) = delete;
-    abstract_operator(abstract_operator&&) = default;
+ public:
+  abstract_operator(const std::shared_ptr<abstract_operator> left = nullptr,
+                    const std::shared_ptr<abstract_operator> right = nullptr);
+  abstract_operator(abstract_operator const&) = delete;
+  abstract_operator(abstract_operator&&) = default;
 
-    virtual void execute() = 0;
-    virtual std::shared_ptr<table> get_output() const = 0;
+  virtual void execute() = 0;
+  virtual std::shared_ptr<table> get_output() const = 0;
 
-protected:
-    virtual const std::string get_name() const = 0;
-    virtual short get_num_in_tables() const = 0;
-    virtual short get_num_out_tables() const = 0;
+ protected:
+  virtual const std::string get_name() const = 0;
+  virtual uint8_t get_num_in_tables() const = 0;
+  virtual uint8_t get_num_out_tables() const = 0;
 
-    const std::shared_ptr<table> _input_left, _input_right;
+  const std::shared_ptr<table> _input_left, _input_right;
 };
 
 class abstract_operator_impl {
-public:
-    virtual void execute() = 0;
-    virtual std::shared_ptr<table> get_output() const = 0;
+ public:
+  virtual void execute() = 0;
+  virtual std::shared_ptr<table> get_output() const = 0;
 };
-
-}
+}  // namespace opossum
